Prep for 1.0 alpha, adapted to runtime changes in main (#31)
### Motivation

On main, the HTTPBody length type changed from Int to Int64.

### Modifications

Adapted the transport with this change.

### Result

Repo builds again when using the latest runtime.

### Test Plan

Adapted tests.

+++ Tests/OpenAPIURLSessionTests/URLSessionBidirectionalStreamingTests/HTTPBodyOutputStreamTests.swift
@@ -30,7 +30,11 @@ class HTTPBodyOutputStreamBridgeTests: XCTestCase {
         let requestBytes = (0...numBytes).map { UInt8($0) }
         let requestChunks = requestBytes.chunks(of: chunkSize)
         let requestByteSequence = MockAsyncSequence(elementsToVend: requestChunks, gatingProduction: false)
-        let requestBody = HTTPBody(requestByteSequence, length: .known(requestBytes.count), iterationBehavior: .single)
+        let requestBody = HTTPBody(
+            requestByteSequence,
+            length: .known(Int64(requestBytes.count)),
+            iterationBehavior: .single
+        )
 
         // Create a pair of bound streams with a tiny buffer to be the bottleneck for backpressure.
         var inputStream: InputStream?
@@ -77,7 +81,11 @@ class HTTPBodyOutputStreamBridgeTests: XCTestCase {
         let requestBytes = (0...numBytes).map { UInt8($0) }
         let requestChunks = requestBytes.chunks(of: chunkSize)
         let requestByteSequence = MockAsyncSequence(elementsToVend: requestChunks, gatingProduction: true)
-        let requestBody = HTTPBody(requestByteSequence, length: .known(requestBytes.count), iterationBehavior: .single)
+        let requestBody = HTTPBody(
+            requestByteSequence,
+            length: .known(Int64(requestBytes.count)),
+            iterationBehavior: .single
+        )
 
         // Create a pair of bound streams with a tiny buffer to be the bottleneck for backpressure.
         var inputStream: InputStream?
@@ -129,7 +137,11 @@ class HTTPBodyOutputStreamBridgeTests: XCTestCase {
         let requestBytes = (0...numBytes).map { UInt8($0) }
         let requestChunks = requestBytes.chunks(of: chunkSize)
         let requestByteSequence = MockAsyncSequence(elementsToVend: requestChunks, gatingProduction: true)
-        let requestBody = HTTPBody(requestByteSequence, length: .known(requestBytes.count), iterationBehavior: .single)
+        let requestBody = HTTPBody(
+            requestByteSequence,
+            length: .known(Int64(requestBytes.count)),
+            iterationBehavior: .single
+        )
 
         // Create a pair of bound streams with a tiny buffer to be the bottleneck for backpressure.
         var inputStream: InputStream?
@@ -183,7 +195,11 @@ class HTTPBodyOutputStreamBridgeTests: XCTestCase {
         let requestBytes = (0...numBytes).map { UInt8($0) }
         let requestChunks = requestBytes.chunks(of: chunkSize)
         let requestByteSequence = MockAsyncSequence(elementsToVend: requestChunks, gatingProduction: true)
-        let requestBody = HTTPBody(requestByteSequence, length: .known(requestBytes.count), iterationBehavior: .single)
+        let requestBody = HTTPBody(
+            requestByteSequence,
+            length: .known(Int64(requestBytes.count)),
+            iterationBehavior: .single
+        )
 
         // Create a pair of bound streams with a tiny buffer to be the bottleneck for backpressure.
         var inputStream: InputStream?
@@ -240,7 +256,11 @@ class HTTPBodyOutputStreamBridgeTests: XCTestCase {
         let requestBytes = (0...numBytes).map { UInt8($0) }
         let requestChunks = requestBytes.chunks(of: chunkSize)
         let requestByteSequence = MockAsyncSequence(elementsToVend: requestChunks, gatingProduction: true)
-        let requestBody = HTTPBody(requestByteSequence, length: .known(requestBytes.count), iterationBehavior: .single)
+        let requestBody = HTTPBody(
+            requestByteSequence,
+            length: .known(Int64(requestBytes.count)),
+            iterationBehavior: .single
+        )
 
         // Create a pair of bound streams with a tiny buffer to be the bottleneck for backpressure.
         var inputStream: InputStream?
